Major updates to nxrefine
Includes new function - copy_parameters - to transfer alignment
information between entries and files, and moves orientation_matrix to
the detector group, where it is now defined as the U matrix, rather
than the UB matrix. This allows the NXsample group to be the same
(i.e., linked) in all the entries.