Prevent unnecessary creation of tasks directory
This only initializes the logger, server, and database when needed. If the file used to create NXReduce is not in the proper file structure, these will create a redundant tasks directory in an apparently arbitrary location. This will still happen if the user initiates a workflow operation, but they can be instructed never to do this unless the file is in the right location.