Configure queues to allow regular and masked operations
A new `regular` switch can be used in conjunction with the `mask` switch
to allow both operations to be queued at the same time. This is used in
`Manage Workflows` dialog to allow all tasks to be submitted
simultaneously. The new configuration also ensures that all tasks added
to the queue from the command line pass all the relevant parameters.